Enable boundary movement with prescribed position
This enables ALE mesh boundary movement with a prescribed discrete
function interpreted as a position in time.

This feature also requires updating the mesh velocity BCs in every time
step stage (as there can now be a prescribed inhomogeneous Dirichlet BC
on moving boundaries).

This requires being able to specify Dirichlet BCs on mesh velocity by
skipping updates to Dirichlet BC nodes during mesh the velocity
smoother. This is done by skipping those nodes during CSR::mult, which
changes the baselines of both regression tests that exercise this
smoothing. The change in the diagnostics is not detectable and the
change in the field output is minor, visually inspected, and appears
correct.