Move querying meshvel symbc nodes to ALE
This commit moves the code that queries the nodes (and associated side
set ids) at which ALE mesh velocity symmetry BCs are set to ALE from
ALECG. This was a leftover that I forgot from the last set of commits.
This code should be in ALE, together with the rest of the ALE-related
code.

One slight change in the code, as it is moved, is that the querying code
interrogates boundary nodes instead of boundary faces, which may fix an
issue I'm seeing with a case. Also, this allows reusing this code from
ALE::meshvelBnd, in addition to calling from ALECG::norm(), and thus
makes this, now single, code consistent between what is used to apply
mesh velocity symmetry BCs and what is used to compute boundary point
normals.