Ensure boundary faces are pruned
Previously, if bc_set was true returning from Transporter::matchBCs()
processing bnodes, bface was not processes, which is not what we want,
because then some sidesets that are in the mesh file but not used in the
input file are not deleted, causing divergence in the worst case. This
commit fixes that by switching the order the return value is used from
matchBCs(). This will break some regression tests, which will be fixed
next.