Separate sidesets used into physics and ALE
This commit separates sidesets and their associated node lists into what
is needed for physics boundary conditions and those used for ALE, i.e.,
mesh velocity boundary conditions. This is necessary because even if ALE
is not used, if sidesets are present in the input file and so passed
into the physics class, e.g., ALECG, the bface and bnode data
structures will have these nodes and may compute different normals if
some nodes are shared with sidesets that share mesh nodes belonging to
sidesets that are used for ALE as well as physics BCs. In a worst case,
this may lead to a blow up due to the wrong normals.

This commit puts in code to separate the sidesets used for physics and
for ALE and use the respective data structures for their purpose only.

Note that the separation of the sidesets not in Transporter because
Transporter::matchBCs only filters the sidesets in the mesh file for
what is used to configure the problem by the user. This way the bface
and bnode data are pushed through global reordering, inital AMR, etc.
Separation happens in Discretization because that way not only ALECG can
use it but also other schemes that need ALE in the future.

One regression test had a minor update, only in the field output, not in
the diagnostics (integral) variables, and I inspected the solution of
the test (Sod with ALE without any smoothing), running out a lot longer
in time and the solution appears correct as before.