bugfix: support integration with ExaM2M transfer every iteration
This revised the callback scheme to be compatible with recent
changes to ExaM2M.  ExaM2M will trigger a single callback to
the destination mesh when the solution transfer is complete.
Discretization will set that callback to be transferCompleteCBinDest.
That will call the callback passed in by the destination mesh solver
in the transfer call, and call transferCompleteNotifyFromDest in the
Discretization source mesh, which will call the callback passed in
by the source mesh solver.