Geometry-based hole search
This commit is an initial implementation of hole search for overset. The
implementation is roughly correct, but it does not work, so there must
be a mistake somewhere. However, we abandon this idea in favor of an
integral-based algorithm. Below are some notes regarding this
geometry-based hole search in case this is interesting in the future for
some reason.

The idea of the geometry-based hole search is to decide whether a mesh
node is inside of a hole or not based on two conditions:

(1) The sign of the scalar product of the vector connecting the node
with a point of a surface triangle and the normal vector of the surface
triangle is the same sign for hole points for all surface triangles.
This works for convex holes but not for concave ones. Concave holes need
the 2nd condition below to be also satisfied.

(2) The scalar product for a surface triangle is ignored if the mesh
node is not visible to a surface triangle. Visibility is defined by a
given surface triangle being closest to the mesh node. That is, if a line
connecting the mesh node with a point of surface triangle intersects
another surface triangle, the former triangle is invisible to the point.

To decide whether the node is visible to a surface triangle the
following algorithm is implemented, see also

https://stackoverflow.com/questions/42740765/intersection-between-line-and-triangle-in-3d

This algorithm requires a nested double-loop over all surface triangles
for each mesh node. As a result, even this worked, it would have M * N^2
complexity where M is the number of mesh nodes and N is the number of
surface triangles.

for every mesh point i

  let q2 be mesh point i

  for every surface triangle t

    let q1 be a point of surface triangle t

    for every surface triangle j

      let p1, p2, and p3 denote the points of surface triangle j

      let vol(a,b,c,d) be the sign of tetrahedron volume given by points
      a, b, c, and d

      let q = vol(q1,q2,p1,p2)

      if none of
         (vol(q1,p1,p2,p3) != vol(q2,p1,p2,p3)) &&
          q == vol(q1,q2,p2,p3) && q == vol(q1,q2,p3,p1) for each j
      then i is visible to t

    end
  end
end

The inner j loop is implemented using std::none_of().