MP: Update to allow multiply polynomials for defined temp ranges
New format allows for temperature ranges to be defined in variable
t_range, which expects 4 temperatures (to define 3 ranges). All
chemical species which I expect to be included in our study require
this many ranges.

The cp_coeff structure is updated to be a nested vector, to contain
3 sets of polynomials coefficients per species.

Regression test lua file is updated with the new format.