Updated tolerances in unit test, which will now pass if solver's exit tolerance is set lower than unit test's expected tolerance
Still not a proper unit test.  Only checks reported residual from
CG solver.  Does NOT actually test compared to known solution. In
short, it checks that the solver thinks it is solved to a certain
tolerance, not that it actually is.  Which is the only reason it
was previously passing...