Working version that solves the unit tests (also updated) and the sod problem.
Updated logic with small numbers, but this is not yet satisfactory.
Fixed some minor bugs and hopefully has clear logic.  New routine
added to check actual tolerance at end, which was not done
correctly before.

Still needs clean-up, but getting closer.  Now fails a different
regression test, but I think this is due to it actually working
where it did not before.